Add nullptr guards to object actions
* Using shell mode, it's possible to invoke object actions without a
  valid document, selection object in place, return early in such cases.

* This is just a hack, the ideal fix would be to make the actions return
  a value in case of unsuccessfull executions and not invoking the rest
  of the commands in the sequence.

* There should be no need of setting the document while performing
  object actions, removing these calls would help surface the underlying
  problems with the code.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/6001

+++ src/actions/actions-object.cpp
@@ -40,7 +40,7 @@ double stod_finite(std::string const &str)
 void object_trace(Glib::VariantBase const &value, InkscapeApplication *app)
 {
     auto selection = app->get_active_selection();
-    if (selection->isEmpty()) {
+    if (!selection || selection->isEmpty()) {
         show_output("action:object_trace: selection empty!", true);
         return;
     }
@@ -143,7 +143,7 @@ void
 object_remove_attribute(Glib::VariantBase const &value, InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
-    if (selection->isEmpty()) {
+    if (!selection || selection->isEmpty()) {
         show_output("action:object_remove_attribute: selection empty!");
         return;
     }
@@ -160,7 +160,7 @@ void
 object_remove_property(Glib::VariantBase const &value, InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
-    if (selection->isEmpty()) {
+    if (!selection || selection->isEmpty()) {
         show_output("action:object_remove_property: selection empty!");
         return;
     }
@@ -190,7 +190,7 @@ object_set_attribute(const Glib::VariantBase& value, InkscapeApplication *app)
     auto const new_value = argument.substr(comma_position + 1);
 
     auto selection = app->get_active_selection();
-    if (selection->isEmpty()) {
+    if (!selection || selection->isEmpty()) {
         show_output("action:object_set_attribute: selection empty!");
         return;
     }
@@ -219,7 +219,7 @@ object_set_property(const Glib::VariantBase& value, InkscapeApplication *app)
     }
 
     auto selection = app->get_active_selection();
-    if (selection->isEmpty()) {
+    if (!selection || selection->isEmpty()) {
         show_output("action:object_set_property: selection empty!");
         return;
     }
@@ -242,10 +242,9 @@ void
 object_unlink_clones(InkscapeApplication *app)
 {
     auto selection = app->get_active_selection();
-
-    // We should not have to do this!
-    auto document  = app->get_active_document();
-    selection->setDocument(document);
+    if (!selection) {
+        return;
+    }
 
     selection->unlink();
 }
@@ -260,6 +259,9 @@ void
 object_clip_set(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // Object Clip Set
     selection->setMask(true, false, should_remove_original());
@@ -270,6 +272,9 @@ void
 object_clip_set_inverse(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // Object Clip Set Inverse
     selection->setMask(true, false, should_remove_original());
@@ -281,6 +286,9 @@ void
 object_clip_release(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // Object Clip Release
     Inkscape::LivePathEffect::sp_remove_powerclip(app->get_active_selection());
@@ -292,6 +300,10 @@ void
 object_clip_set_group(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
+
     selection->setClipGroup();
     // Undo added in setClipGroup().
 }
@@ -300,6 +312,9 @@ void
 object_mask_set(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // Object Mask Set
     selection->setMask(false, false, should_remove_original());
@@ -310,6 +325,9 @@ void
 object_mask_set_inverse(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // Object Mask Set Inverse
     selection->setMask(false, false, should_remove_original());
@@ -321,6 +339,9 @@ void
 object_mask_release(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // Object Mask Release
     Inkscape::LivePathEffect::sp_remove_powermask(app->get_active_selection());
@@ -332,6 +353,9 @@ void
 object_rotate_90_cw(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // Object Rotate 90
     auto desktop = selection->desktop();
@@ -342,6 +366,9 @@ void
 object_rotate_90_ccw(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // Object Rotate 90 CCW
     auto desktop = selection->desktop();
@@ -352,6 +379,9 @@ void
 object_flip_horizontal(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     Geom::OptRect bbox = selection->visualBounds();
     if (!bbox) {
@@ -375,6 +405,9 @@ void
 object_flip_vertical(InkscapeApplication *app)
 {
     Inkscape::Selection *selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     Geom::OptRect bbox = selection->visualBounds();
     if (!bbox) {
@@ -399,20 +432,26 @@ void
 object_to_path(InkscapeApplication *app)
 {
     auto selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
-    // We should not have to do this!
-    auto document  = app->get_active_document();
-    selection->setDocument(document);
     selection->toCurves(false, Inkscape::Preferences::get()->getBool("/options/clonestocurvesjustunlink/value", true));
 }
 
 void
 object_add_corners_lpe(InkscapeApplication *app) {
     auto selection = app->get_active_selection();
+    if (!selection) {
+        return;
+    }
 
     // We should not have to do this!
     auto document  = app->get_active_document();
-    selection->setDocument(document);
+    if (!document) {
+        return;
+    }
+
     auto items = selection->items_vector();
     selection->clear();
     for (auto i : items) {
@@ -436,10 +475,9 @@ void
 object_stroke_to_path(InkscapeApplication *app)
 {
     auto selection = app->get_active_selection();
-
-    // We should not have to do this!
-    auto document  = app->get_active_document();
-    selection->setDocument(document);
+    if (!selection) {
+        return;
+    }
 
     selection->strokesToPaths();
 }
